backend/pyproject.toml
[project]
name = "cicd"
version = "0.1.0"
description = "No-Code graphical CI/CD pipeline builder"
readme = "README.md"
requires-python = ">=3.12"
authors = ["72mins"]
dependencies = [
    "boto3>=1.36.21",
    "django>=5.1.6",
    "django-cors-headers>=4.7.0",
    "django-crum>=0.7.9",
    "django-filter>=24.3",
    "django-storages>=1.14.5",
    "djangorestframework>=3.15.2",
    "djangorestframework-simplejwt>=5.4.0",
    "drf-spectacular>=0.28.0",
    "psycopg2>=2.9.10",
    "pynvim>=0.5.2",
    "python-dotenv>=1.0.1",
    "ruff>=0.9.6",
]
